+++ History.md
@@ -1,3 +1,11 @@
+0.4.3 / 2014-03-27
+==================
+
+* Fix: float():
+	* dont move offset on invalid float
+	* Infinity is a valid float value
+	* quiet(true).float() does not skip invalid float
+
 0.4.2 / 2012-12-04
 ==================